refactor ast node testing into separate classes
